A quick whack at allowing additional HTTP headers to be passed in to a get() call.
Note that get(), fetch_page() and set_headers() all now have the
ability to accept a hash of arguments as a parameter, allowing us
to "stuff" extra parameters into each one while maintaining
backwards compatibility (see tests).

Special headers specified by :etag and :if_modified_since are
recognized, for whatever that's worth.

I realize this is kind of a fundamental change for such a simple
feature, but it looks like there are some API changes planned for
0.8.0 anyway, so I figured I'd throw this out there and start a
conversation about it.