Introduce a "for_domain" flag to Mechanize::Cookie.
Previously, a cookie without the domain attribute would be erroneously
sent to subdomain hosts, which behavior was a violation of RFC 6265
(#4.1.2.3) that had to be fixed.  GH #159

In terms of compatibility, cookie jars serialized with previous
versions of Mechanize using the save_as() method should be safe, but
since there have been many changes in the cookie related code, it is
strongly recommended that you clear them out for regeneration.