Do not include user information or fragment part from a Referer URI.
It is required by RFC 2616 14.36 that a Referer URI MUST NOT include a
fragment part.

As for the user information part there seems to be no explicit mention
in RFC 2616 nor RFC 3986, but I believe we should just follow what
most major browsers do and remove user information part if any for
security reasons.