Fix form field ordering when backed by Hash nodes
When a form field is created with a Hash as the node instead of a
nokogiri node, the sort order was previously undefined.  This can happen
when a user calls `Mechanize.post` or when a Form is manually created
with Hash backed Fields and the user calls `Mechanize.submit` or
`Mechanize::Form#submit`.

fixes #307