test: accept "UTF-8" as an encoding alias for "UTF8"
As of libxml 2.12.0, the `page.encoding` value inspected by this test
is dependent on how libiconv is built (specifically: which aliases are
supported).

- if alias "UTF8" is defined, then the page encoding will be "UTF-8"
- if alias "UTF8" is not defined, then the page encoding will be "UTF8"

The modified test started failing on Windows and MacOS (where Nokogiri
compiles libiconv and statically links it into the
native (precompiled) gem) but still passed on Linux (where libc
dynamically provides libiconv symbols).