fix: correctly recognize encoding errors in libxml >= 2.12.0
Starting with libxml2 v2.12.0, encoding errors have a message that was
not being detected by Page#encoding_error? resulting in a page that
was parsed using incorrect encoding.

This change updates Page to detect those errors as _encoding errors_
and as a result we get the expected encoding back on the parsed
document.